v6.9.2
Compiler: Added the optional language construct "Has WindBarbs [Below | Above]". See Wiki CDL. This feature is beta.
Fix AirLink: Reading the AirLinkLog required an additional condition for preventing an infinite loop (and running out of memory/swap).
Fix Airlink: The windbarb xAxis on top is now aligned with the xAxis below
Fix MeteoCam: Corrected the Upload for the MeteoCam output when the user does not use it.